[DIET-102-Register-page] signup done i guess
