+++ dyeti-backend/src/main/java/ki/agh/dyeti/dto/request/RegisterRequest.java
@@ -2,6 +2,7 @@ package ki.agh.dyeti.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import ki.agh.dyeti.model.util.Gender;
 import lombok.Data;
 
@@ -18,6 +19,6 @@ public class RegisterRequest {
     @NotBlank(message = "Password is required")
     private String password;
 
-    @NotBlank(message = "Gender is required")
+    @NotNull(message = "Gender is required")
     private Gender gender;
 }
