+++ dyeti-frontend/src/pages/index.ts
@@ -2,5 +2,5 @@ export { default as Home } from './Home/Home.tsx';
 export { default as Plans } from './Plans/Plans.tsx';
 export { default as Products } from './Products/Products.tsx';
 export { default as Themes } from './Themes/Themes.tsx';
-export { default as Login } from './Login/Login.tsx';
-export { default as Signup } from './Signup/Signup.tsx';
+export { default as Login } from './Auth/Login/Login.tsx';
+export { default as Signup } from './Auth/Signup/Signup.tsx';
